+++ android/build.gradle
@@ -1,4 +1,4 @@
-def usercentrics_version = "2.7.0"
+def usercentrics_version = "2.7.1"
 version usercentrics_version
 
 buildscript {
