+++ example/src/screens/Home.tsx
@@ -1,16 +1,13 @@
 import React from 'react';
-import { Button, Image, StyleSheet, View } from 'react-native';
+import { Button, StyleSheet, View } from 'react-native';
 import {
-    BannerLogo,
-    BannerSettings,
-    ButtonLayout,
-    ButtonType,
-    FirstLayerOptions,
-    HeaderImageSettings, LegalLinksSettings, SecondLayerOptions,
-    SectionAlignment,
-    Usercentrics,
+    BannerSettings, FirstLayerOptions, SecondLayerOptions, Usercentrics,
     UsercentricsLayout
 } from '../../../src/index';
+import {
+    customizationExampleOne,
+    customizationExampleTwo
+} from './CustomizationExamples';
 
 export const HomeScreen = ({ navigation }: { navigation: any }) => {
     async function showFirstLayer(options: FirstLayerOptions = defaultOptions) {
@@ -77,107 +74,4 @@ export const HomeScreen = ({ navigation }: { navigation: any }) => {
 
 const defaultOptions: FirstLayerOptions = {
     layout: UsercentricsLayout.popupCenter
-}
-
-const customizationExampleOne: FirstLayerOptions = {
-    bannerSettings: {
-        firstLayerStyleSettings: {
-            cornerRadius: 30.0,
-            title: {
-                textSize: 22.0
-            },
-            message: {
-                textSize: 14.0,
-            },
-            buttonLayout: ButtonLayout.row(
-                [
-                    {
-                        buttonType: ButtonType.acceptAll,
-                        cornerRadius: 30.0
-                    }, {
-                        buttonType: ButtonType.more,
-                        backgroundColorHex: "00000000"
-                    }
-                ]
-            )
-        },
-        generalStyleSettings: {
-            font: {
-                regularFont: "Lora",
-                boldFont: "Lora-Regular_Bold",
-                fontSize: 14.0
-            },
-            textColorHex: "#77ff00",
-            layerBackgroundColorHex: "#fff000",
-            logo: createBannerLogo(),
-            links: LegalLinksSettings.secondLayerOnly,
-            toggleStyleSettings: {
-                activeBackgroundColorHex: "#12e09f",
-                inactiveBackgroundColorHex: "#abdccd",
-                disabledBackgroundColorHex: "#d47e00",
-                activeThumbColorHex: "#957950",
-                inactiveThumbColorHex: "#4439e9",
-                disabledThumbColorHex: "c99dc1"
-            }
-        }
-    },
-    layout: UsercentricsLayout.popupBottom
-}
-
-const customizationExampleTwo: FirstLayerOptions = {
-    layout: UsercentricsLayout.full,
-    bannerSettings: {
-        firstLayerStyleSettings: {
-            cornerRadius: 20,
-            backgroundColorHex: "000000",
-            headerImage: HeaderImageSettings.hidden(),
-            title: {
-                textSize: 22.0,
-                textColorHex: "FFFFFF",
-                textAlignment: SectionAlignment.center
-            },
-            message: {
-                textSize: 18.0,
-                textColorHex: "FFFFFF"
-            },
-            buttonLayout: ButtonLayout.grid([
-                [
-                    {
-                        buttonType: ButtonType.acceptAll,
-                        cornerRadius: 30.0,
-                        backgroundColorHex: "FFFFFF",
-                        textColorHex: "000000",
-                    }, {
-                        buttonType: ButtonType.denyAll,
-                        backgroundColorHex: "00000000",
-                        textColorHex: "FFFFFF"
-                    }
-                ], [
-                    {
-                        buttonType: ButtonType.more,
-                        cornerRadius: 30.0,
-                        textColorHex: "000000"
-                    }
-                ]
-            ])
-        },
-        generalStyleSettings: {
-            font: {
-                regularFont: "Lora",
-                boldFont: "Lora-Regular_Bold",
-                fontSize: 14.0
-            },
-            textColorHex: "#ffffff",
-            layerBackgroundColorHex: "#000000",
-            layerBackgroundSecondaryColorHex: "#6b6b6b",
-            logo: createBannerLogo(),
-            links: LegalLinksSettings.secondLayerOnly
-        }
-    }
-}
-
-function createBannerLogo(): BannerLogo {
-    // Logo name is used for iOS and the Image.resolveAssetSource is used for Android.
-    const customLogo = new BannerLogo("logo.png", Image.resolveAssetSource(require('../../assets/images/logo.png')))
-    return customLogo
 }
