+++ package.json
@@ -1,13 +1,13 @@
 {
     "name": "@usercentrics/react-native-sdk",
-    "version": "2.7.0",
+    "version": "2.7.1",
     "description": "Usercentrics SDK",
     "homepage": "https://usercentrics.com",
     "main": "lib/index.js",
     "types": "lib/index.d.ts",
     "author": "Usercentrics <developer@usercentrics.com>",
     "iosPackageName": "react-native-usercentrics",
-    "iosPackageVersion": "2.7.0",
+    "iosPackageVersion": "2.7.1",
     "license": "SEE LICENSE IN LICENSE",
     "files": [
         "android",
