+++ src/models/TCF2Settings.tsx
@@ -1,110 +1,108 @@
-export class TCF2Settings { 
-    firstLayerTitle: string
-    secondLayerTitle: string
-    tabsPurposeLabel: string
-    tabsVendorsLabel: string
-    labelsFeatures: string
-    labelsIabVendors: string
-    labelsNonIabPurposes: string
-    labelsNonIabVendors: string
-    labelsPurposes: string
-    vendorFeatures: string
-    vendorLegitimateInterestPurposes: string
-    vendorPurpose: string
-    vendorSpecialFeatures: string
-    vendorSpecialPurposes: string
-    togglesConsentToggleLabel: string
-    togglesLegIntToggleLabel: string
-    buttonsAcceptAllLabel: string
-    buttonsDenyAllLabel: string
-    buttonsSaveLabel: string
-    linksManageSettingsLabel: string
-    linksVendorListLinkLabel: string
-    cmpId: number
-    cmpVersion: number
-    firstLayerHideToggles: boolean
-    secondLayerHideToggles: boolean
-    hideLegitimateInterestToggles: boolean
-    firstLayerHideButtonDeny?: boolean
-    secondLayerHideButtonDeny: boolean
-    publisherCountryCode: string
-    purposeOneTreatment: boolean
-    selectedVendorIds: [number]
-    gdprApplies: boolean
-    consensuDomain: string
-    consensuScriptPath: string
-    selectedStacks: [number]
-    disabledSpecialFeatures: [number]
-    firstLayerShowDescriptions: boolean
-    hideNonIabOnFirstLayer: boolean
-    resurfaceIABLegalBasisChanged: boolean
-    resurfacePeriodEnded: boolean
-    resurfacePurposeChanged: boolean
-    resurfaceVendorAdded: boolean
-    vendorToggleAll: boolean
-    firstLayerDescription: string
-    firstLayerAdditionalInfo: string
-    secondLayerDescription: string
-    togglesSpecialFeaturesToggleOn: string
-    togglesSpecialFeaturesToggleOff: string
-    appLayerNoteResurface: string
-    firstLayerNoteGlobal: string
-    firstLayerNoteResurface: string
+export class TCF2Settings {
+  firstLayerTitle: string
+  secondLayerTitle: string
+  tabsPurposeLabel: string
+  tabsVendorsLabel: string
+  labelsFeatures: string
+  labelsIabVendors: string
+  labelsNonIabPurposes: string
+  labelsNonIabVendors: string
+  labelsPurposes: string
+  vendorFeatures: string
+  vendorLegitimateInterestPurposes: string
+  vendorPurpose: string
+  vendorSpecialFeatures: string
+  vendorSpecialPurposes: string
+  togglesConsentToggleLabel: string
+  togglesLegIntToggleLabel: string
+  buttonsAcceptAllLabel: string
+  buttonsDenyAllLabel: string
+  buttonsSaveLabel: string
+  linksManageSettingsLabel: string
+  linksVendorListLinkLabel: string
+  cmpId: number
+  cmpVersion: number
+  firstLayerHideToggles: boolean
+  secondLayerHideToggles: boolean
+  hideLegitimateInterestToggles: boolean
+  firstLayerHideButtonDeny?: boolean
+  secondLayerHideButtonDeny: boolean
+  publisherCountryCode: string
+  purposeOneTreatment: boolean
+  selectedVendorIds: [number]
+  gdprApplies: boolean
+  consensuDomain: string
+  consensuScriptPath: string
+  selectedStacks: [number]
+  disabledSpecialFeatures: [number]
+  firstLayerShowDescriptions: boolean
+  hideNonIabOnFirstLayer: boolean
+  resurfacePeriodEnded: boolean
+  resurfacePurposeChanged: boolean
+  resurfaceVendorAdded: boolean
+  vendorToggleAll: boolean
+  firstLayerDescription: string
+  firstLayerAdditionalInfo: string
+  secondLayerDescription: string
+  togglesSpecialFeaturesToggleOn: string
+  togglesSpecialFeaturesToggleOff: string
+  appLayerNoteResurface: string
+  firstLayerNoteGlobal: string
+  firstLayerNoteResurface: string
 
 
   constructor(
-    firstLayerTitle: string, 
-    secondLayerTitle: string, 
-    tabsPurposeLabel: string, 
-    tabsVendorsLabel: string, 
-    labelsFeatures: string, 
-    labelsIabVendors: string, 
-    labelsNonIabPurposes: string, 
-    labelsNonIabVendors: string, 
-    labelsPurposes: string, 
-    vendorFeatures: string, 
-    vendorLegitimateInterestPurposes: string, 
-    vendorPurpose: string, 
-    vendorSpecialFeatures: string, 
-    vendorSpecialPurposes: string, 
-    togglesConsentToggleLabel: string, 
-    togglesLegIntToggleLabel: string, 
-    buttonsAcceptAllLabel: string, 
-    buttonsDenyAllLabel: string, 
-    buttonsSaveLabel: string, 
-    linksManageSettingsLabel: string, 
-    linksVendorListLinkLabel: string, 
-    cmpId: number, 
-    cmpVersion: number, 
-    firstLayerHideToggles: boolean, 
-    secondLayerHideToggles: boolean, 
-    hideLegitimateInterestToggles: boolean, 
-    secondLayerHideButtonDeny: boolean, 
-    publisherCountryCode: string, 
-    purposeOneTreatment: boolean, 
-    selectedVendorIds: [number], 
-    gdprApplies: boolean, 
-    consensuDomain: string, 
-    consensuScriptPath: string, 
-    selectedStacks: [number], 
-    disabledSpecialFeatures: [number], 
-    firstLayerShowDescriptions: boolean, 
-    hideNonIabOnFirstLayer: boolean, 
-    resurfaceIABLegalBasisChanged: boolean, 
-    resurfacePeriodEnded: boolean, 
-    resurfacePurposeChanged: boolean, 
-    resurfaceVendorAdded: boolean, 
-    vendorToggleAll: boolean, 
-    firstLayerDescription: string, 
-    firstLayerAdditionalInfo: string, 
-    secondLayerDescription: string, 
-    togglesSpecialFeaturesToggleOn: string, 
-    togglesSpecialFeaturesToggleOff: string, 
-    appLayerNoteResurface: string, 
-    firstLayerNoteGlobal: string, 
+    firstLayerTitle: string,
+    secondLayerTitle: string,
+    tabsPurposeLabel: string,
+    tabsVendorsLabel: string,
+    labelsFeatures: string,
+    labelsIabVendors: string,
+    labelsNonIabPurposes: string,
+    labelsNonIabVendors: string,
+    labelsPurposes: string,
+    vendorFeatures: string,
+    vendorLegitimateInterestPurposes: string,
+    vendorPurpose: string,
+    vendorSpecialFeatures: string,
+    vendorSpecialPurposes: string,
+    togglesConsentToggleLabel: string,
+    togglesLegIntToggleLabel: string,
+    buttonsAcceptAllLabel: string,
+    buttonsDenyAllLabel: string,
+    buttonsSaveLabel: string,
+    linksManageSettingsLabel: string,
+    linksVendorListLinkLabel: string,
+    cmpId: number,
+    cmpVersion: number,
+    firstLayerHideToggles: boolean,
+    secondLayerHideToggles: boolean,
+    hideLegitimateInterestToggles: boolean,
+    secondLayerHideButtonDeny: boolean,
+    publisherCountryCode: string,
+    purposeOneTreatment: boolean,
+    selectedVendorIds: [number],
+    gdprApplies: boolean,
+    consensuDomain: string,
+    consensuScriptPath: string,
+    selectedStacks: [number],
+    disabledSpecialFeatures: [number],
+    firstLayerShowDescriptions: boolean,
+    hideNonIabOnFirstLayer: boolean,
+    resurfacePeriodEnded: boolean,
+    resurfacePurposeChanged: boolean,
+    resurfaceVendorAdded: boolean,
+    vendorToggleAll: boolean,
+    firstLayerDescription: string,
+    firstLayerAdditionalInfo: string,
+    secondLayerDescription: string,
+    togglesSpecialFeaturesToggleOn: string,
+    togglesSpecialFeaturesToggleOff: string,
+    appLayerNoteResurface: string,
+    firstLayerNoteGlobal: string,
     firstLayerNoteResurface: string,
-    firstLayerHideButtonDeny?: boolean, 
-) {
+    firstLayerHideButtonDeny?: boolean,
+  ) {
     this.firstLayerTitle = firstLayerTitle
     this.secondLayerTitle = secondLayerTitle
     this.tabsPurposeLabel = tabsPurposeLabel
@@ -143,7 +141,6 @@ export class TCF2Settings {
     this.disabledSpecialFeatures = disabledSpecialFeatures
     this.firstLayerShowDescriptions = firstLayerShowDescriptions
     this.hideNonIabOnFirstLayer = hideNonIabOnFirstLayer
-    this.resurfaceIABLegalBasisChanged = resurfaceIABLegalBasisChanged
     this.resurfacePeriodEnded = resurfacePeriodEnded
     this.resurfacePurposeChanged = resurfacePurposeChanged
     this.resurfaceVendorAdded = resurfaceVendorAdded
